requirements.txt
groq
python-dotenv
pytest

